pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cleo"
version = "0.1.0"
description = "Cleo — AI-powered smart glasses platform"
requires-python = ">=3.10,<3.13"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "faiss-cpu>=1.7.4",
    "nemo-toolkit[asr]>=2.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "PyAudio>=0.2.14",
    "opencv-python>=4.8.0",
    "soundfile>=0.12.0",
    "structlog>=23.1.0",
    "boto3>=1.34.0",
    "pytest>=8.0.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-mock>=3.12"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["services", "apps", "generated"]
